feat: Dropdown 컴포넌트 TS로 변경, attribute 활용 부분을 전체 데이터가 아닌 sort, category만
